Added notification tests
